Made some calculations
Added all the trades, and started working on how food is made. Idk I'm probably making a stupid mistake on the food calculation. I want the first 1000 farmers to make 5.1 units per and then the rest start to make less, but I think I'm making a stupid mistake because the way the loop is set up rn it should always have positive food but it's not.
w/e I'm tired. lol